wb7: workaround: ignore power loss in PCF85063
During power-up, oscillation is disturbed by EMI from DC-DC.
It results in a couple of missing oscillation cycles and thus have a negligible
effect on clock accuracy.
Unfortunately, it will also result in power loss bit being set in RTC and RTC
data marked as invalid in kernel.

So, ignore this bit.