ARM: dts: r40: Add dts for Wiren Board 7.4
Wiren Board 7.4 is a building automation controller with a lot
of wired and wireless connectivity options.

It's based on Allwinner A40i SoC, which is industrial-grade
version of R40.

The main difference from Wiren Board 7.3 or older is the Embedded
Controller - a chip with such function as RTC, ADC, GPIO, watchdog
and power control