mdio-gpio: add external clock support for WB8 EPHY
This is necessary for Wiren Board 8 external PHY
which is connected via mdio-gpio but uses EPHY_25M clock
from Allwinner T507 chip.

This clock must be enabled before MDIO bus scan. I tried
to achieve it using other DTS tricks (regulators with clocks,
parent devices with clocks) but didn't succeed.
Maybe there is a proper working way to do it though.

Not sure that this patch will be accepted in upstream.